Added Boss Ship. (#34)
Added simple models for components. Created AI for the boss. Modified
the boss explosion to look a little better. Created prefab for the boss
ship(colliders, turrets etc) and added it to the unit references.

Still to be done: Boss health meter and better message boxes. Need to
have the level for the boss ready before these can be done properly
though.